requirements.txt
agno
openai
firecrawl-py==1.9.0
gradio==5.9.1
pydantic
dataclasses